+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file. The format 
 
 
 ## Unreleased
+### Improved
+- Radio button spacing and indicator alignment
 ## [v0.8.1]
 ### Improved
 - Accordion chevron orientation and animation performance